requirements.txt
streamlit>=1.0
python-dotenv>=0.21.0
openai>=0.27.0